+++ src/Model/Serializer/ValueObject/Calendar/CalendarSerializer.php
@@ -5,10 +5,10 @@ declare(strict_types=1);
 namespace CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar;
 
 use Broadway\Serializer\Serializable;
-use CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar\OpenHours\AdjustedDaysDenormalizer;
-use CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar\OpenHours\AdjustedDayNormalizer;
-use CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar\OpenHours\ClosedDayNormalizer;
-use CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar\OpenHours\ClosedDaysDenormalizer;
+use CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar\OpeningHours\AdjustedDaysDenormalizer;
+use CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar\OpeningHours\AdjustedDayNormalizer;
+use CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar\OpeningHours\ClosedDayNormalizer;
+use CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar\OpeningHours\ClosedDaysDenormalizer;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\BookingAvailability;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\BookingAvailabilityType;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\Calendar;
